requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
sqlite3-binary
duckduckgo-search
